+++ internal/constant/types/manager.go
@@ -13,7 +13,7 @@ type CreateResponse struct {
 	Id        int       `json:"id"`
 	Ident     string    `json:"ident"`
 	Name      string    `json:"name"`
-	Key       uint32    `json:"key"`
+	Key       uint64    `json:"key"`
 	Secret    string    `json:"secret"`
 	Status    int8      `json:"status"`
 	ExpiredAt time.Time `json:"expired_at"`
@@ -23,7 +23,7 @@ type CreateResponse struct {
 type CreateData struct {
 	Ident     string    `json:"ident"`
 	Name      string    `json:"name"`
-	Key       uint32    `json:"key"`
+	Key       uint64    `json:"key"`
 	Secret    string    `json:"secret"`
 	Status    int8      `json:"status"`
 	ExpiredAt time.Time `json:"expired_at"`
